+++ packages/vulcan-admin/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'vulcan:admin',
   summary: 'Vulcan components package',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/VulcanJS/Vulcan.git',
 });
 
@@ -12,7 +12,7 @@ Package.onUse(function(api) {
     'fourseven:scss@4.12.0',
     'dynamic-import@0.1.1',
     // Vulcan packages
-    'vulcan:core@=1.15.1',
+    'vulcan:core@=1.15.2',
   ]);
 
   api.mainModule('lib/server/main.js', 'server');