+++ packages/vulcan-backoffice/package.js
@@ -1,12 +1,12 @@
 Package.describe({
   name: 'vulcan:backoffice',
   summary: 'Vulcan automated backoffice generator',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/VulcanJS/Vulcan.git',
 });
 
 Package.onUse(api => {
-  api.use(['vulcan:core@=1.15.1', 'vulcan:i18n@=1.15.1', 'vulcan:accounts@1.15.1']);
+  api.use(['vulcan:core@=1.15.2', 'vulcan:i18n@=1.15.2', 'vulcan:accounts@1.15.2']);
 
   api.mainModule('lib/server/main.js', 'server');
   api.mainModule('lib/client/main.js', 'client');