+++ packages/vulcan-email/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: 'vulcan:email',
   summary: 'Vulcan email package',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/VulcanJS/Vulcan.git',
 });
 
 Package.onUse(function(api) {
   api.versionsFrom('1.6.1');
 
-  api.use(['vulcan:lib@=1.15.1']);
+  api.use(['vulcan:lib@=1.15.2']);
 
   api.mainModule('lib/server.js', 'server');
   api.mainModule('lib/client.js', 'client');