+++ packages/vulcan-forms-tags/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: 'vulcan:forms-tags',
   summary: 'Vulcan tag input package',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/VulcanJS/Vulcan.git',
 });
 
 Package.onUse(function(api) {
   api.versionsFrom('1.6.1');
 
-  api.use(['vulcan:core@=1.15.1', 'vulcan:forms@=1.15.1']);
+  api.use(['vulcan:core@=1.15.2', 'vulcan:forms@=1.15.2']);
 
   api.mainModule('lib/export.js', ['client', 'server']);
 });