+++ packages/vulcan-forms/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: 'vulcan:forms',
   summary: 'Form containers for React',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/meteor-utilities/react-form-containers.git',
 });
 
 Package.onUse(function(api) {
   api.versionsFrom('1.6.1');
 
-  api.use(['vulcan:core@=1.15.1']);
+  api.use(['vulcan:core@=1.15.2']);
 
   api.mainModule('lib/client/main.js', ['client']);
   api.mainModule('lib/server/main.js', ['server']);