+++ packages/vulcan-lib/lib/modules/config.js
@@ -9,7 +9,7 @@ import SimpleSchema from 'simpl-schema';
 Vulcan = {};
 
 // eslint-disable-next-line no-undef
-Vulcan.VERSION = '1.15.1';
+Vulcan.VERSION = '1.15.2';
 
 // ------------------------------------- Schemas -------------------------------- //