+++ packages/vulcan-voting/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: 'vulcan:voting',
   summary: 'Vulcan scoring package.',
-  version: '1.15.1',
+  version: '1.15.2',
   git: 'https://github.com/VulcanJS/Vulcan.git',
 });
 
 Package.onUse(function(api) {
   api.versionsFrom('1.6.1');
 
-  api.use(['fourseven:scss@4.12.0', 'vulcan:core@=1.15.1', 'vulcan:i18n@=1.15.1'], ['client', 'server']);
+  api.use(['fourseven:scss@4.12.0', 'vulcan:core@=1.15.2', 'vulcan:i18n@=1.15.2'], ['client', 'server']);
 
   api.mainModule('lib/server/main.js', 'server');
   api.mainModule('lib/client/main.js', 'client');