+++ packages/vulcan-ui-material/history.md
@@ -27,7 +27,7 @@
 1.13.0_1 / 2019-07-23
 =====================
 
- * TooltipButton: Deprecated TooltipIntl and TooltipIconButton in favor of TooltipButton - they will be deleted in Vulcan 1.15.1
+ * TooltipButton: Deprecated TooltipIntl and TooltipIconButton in favor of TooltipButton - they will be deleted in Vulcan 1.15.2
  
 1.13.0 / 2019-07-19
 ===================